BackDB/requirements.txt
alembic==1.3.2
APScheduler==3.6.3
bcrypt==3.1.7
Bootstrap-Flask==1.2.0
cffi==1.13.2
Click==7.0
cryptography==2.8
dominate==2.4.0
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-SSE==0.2.1
Flask-WTF==0.14.2
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
paramiko==2.7.1
pycparser==2.19
pycryptodome==3.9.4
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.11.1
python-socketio==4.4.0
pytz==2019.3
redis==3.3.11
six==1.13.0
SQLAlchemy==1.3.12
SQLAlchemy-Utils==0.36.1
tornado==6.0.3
tzlocal==2.0.0
visitor==0.1.3
webssh==1.5.0
Werkzeug==0.16.0
WTForms==2.2.1
